backend/requirements.txt
# Core
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-multipart

# Database
asyncpg
redis

# Auth
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1

# Utilities
python-dotenv
aiohttp
loguru
